+++ jest-setup.js
@@ -1,7 +1,1 @@
-import { configure } from 'enzyme';
-import Adapter from 'enzyme-adapter-react-16';
-
-/**
- * Configure for React
- */
-configure({ adapter: new Adapter() });
+import './.config/jest-setup';