+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 
 - Update to Grafana 9.5.2 (#70)
 - Add E2E Cypress testing (#71)
+- Migrate to Plugin Tools 1.5.0 (#72)
+- Update to Node 18 (#72)
+- Increase Test Coverage (#72)
 
 ## 3.5.0 (2023-03-07)